Prevent exception on caps when doing stems for unknown
With DO STEMS FOR UKNOWN = Y and IGNORE CAPS = N, a word with a capital letter will raise an exception.  Similar result if IGNORE NAMES = N and the input has a capitalized first letter.

Converting the unknown word to all lower case prevents that exception.

This corrects what appears to be an oversight in the original code - cf. a few lines below in the function where the S constant is converted to lower case.