Format examples line consistently
This fixes an inconsistency between the formatting of DAT and ABL examples vs. all others.

The DAT and ABL cases skip a line before outputting the example line.  The comment in Whitaker's original code says that's because the examples lines are "too much for same line."

That's incorrect.  See under the VPAR case below, where there are examples lines longer than the DAT and ABL examples lines.