Fix case ADJ 3 4 (should be case ADJ 3 6)
This should resolve issue #105.

When dictionary_form's conditionals were converted to case statements the "6" in ADJ case 3, varient 6 was made a "4".  There is no variant 4, but there is a variant 6.

Makeewds calls dictionary_form to build EWDSLIST.GEN.  When it hits the first ADJ 3 6 (stem 1:  amethystizon-, line 3003) makeewds raises an exception and exits.

This should also correct the output of ADJ 3 6 in Latin mode.

Of course you need to follow the instructions to rebuild your English dictionary files after applying this patch.