Prevent dupes in English output
Stop dupes in English output.  An example is "utor" in issue #105.

Previously, if a word appeared twice in a definition line, that word would be coded as two separate English word entries => duplicate output.

This diff adds a loop to remove dupes in the MAKEEWDS program. The first appearance of a word is kept, which matters for the entry's "RANK".

[Note: This removes some dupes that running ONERS as described in Whitaker's original HOWTO does not eliminate]