Fix hang at end of file when COMMAND_LINE_FILE is true
When words is run from the command line with input and output files as parameters it currently hangs, waiting for two returns just as if it is in interactive mode.  This happens because process_input re-sets to standard input and then closes the input file.

To fix this cleanly - without relying on GIVE_UP exception - just return to the main calling procedure.  There's nothing left to do in that procedure except a conditional that we can never satisfy, so the return runs into the end of the program.