Issue 101 - diff 2 of 4 - INFLECTS
Change 2 7 to conform to A&G 82, 52 to address issue #101

This change breaks one 2 7 entry:  echo.  The echo entry was incorrect, so I have conformed it to OED and L+S.  It otherwise comports with the Latin 4th, which appears to be the view of those dictionaries.  I haven't yet found anything indicates this approach is incorrect.