Create build_binaries.yml
Added github action to make, run tests, and export binaries for macOS and linux.

I got around #109 with a bit of a hack. This does not solve the problem, but it allows us to get it working for now.

The same functionality is half there for Windows, I just haven't figured out how to properly install GNAT yet.