Allow selection of the search path for input files
An environment variable allows to select datadir at run time. This is
useful for building and testing dictionaries.

A Make variable allows to select the default datadir at build
time. This allows redistributors to conform to their local conventios
(/usr/local for example).

If none is set, the behaviour is unchanged.